requirements.txt
python-jenkins==1.7.0
lxml==4.5.0